% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_population.R
\docType{data}
\name{ref_population}
\alias{ref_population}
\title{European Standard Population 2013 (abridged)}
\format{
A data frame with 19 rows and 2 variables:
\describe{
\item{AgeBand}{Age group (character, e.g. "0-4", "5-9")}
\item{StdPopulation}{Standard population count for that age band}
}
}
\source{
Eurostat ESP 2013
}
\usage{
ref_population
}
\description{
A dataset containing abridged ESP 2013 standard population counts
for use in age standardisation.
}
\keyword{datasets}
